Add ability to seed multiple tables
Added some new tables with foreign keys to prove it works.

We need quite a few scripts to make this work, and some manual
faffing with the JSON files describing the data.

I think Synth assumes it's looking at a full database, so if
we work on an empty one then it will generate empty data by
default. So we need to twiddle with some values.